requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
requests==2.31.0
websockets==12.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
shapely==2.0.2
geojson==3.1.0
haversine==2.8.0
python-dotenv==1.0.0